Use a QMacCocoaViewContainer for VlcVideoWidget on Mac OS X
This fixes video playback in non-top-level widgets on Mac OS X, using
libvlc_media_player_set_nsobject() instead of *_player_set_agl(). The
VlcVideoWidget is now a QMacCocoaViewContainer (on Mac OS X only), which
was added in Qt 4.6. Instead of _widget, it wraps a standard NSView
which will be used for video playback.

Since libvlc_media_player_set_nsobject() requires an void pointer
instead of a WId, a new typedef named WindowId has been added. It is
a void pointer on Mac OS X and a WId on all other systems, so the rest
of the library can work like before.

This solution required a bit of Objective-C in VideoWidget.cpp, so the
necessary compiler flags have been added to CMakeLists.txt.